added Rich Text Editor
1. import tinyMCE-react for provide Rich Text Editor in upload/edit page
2. developed some components and added comment to increase readability of code

+++ client/components/settingPage/userProfile_components/EditPassword.tsx
@@ -1,5 +1,6 @@
 import React, { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
+import { validateField } from '../../ManageFormData';
 
 interface EditPasswordComponentProps {
     updateSection: (section: string) => void;
@@ -13,6 +14,7 @@ const EditPasswordComponent = ({updateSection}: EditPasswordComponentProps) => {
 
     const navigate = useNavigate();
 
+    // Keep the user input during edit
     const OnInputChange = (event: React.ChangeEvent<HTMLInputElement>) => {
         const {name, value} = event.target;
         switch (name) {
@@ -25,32 +27,40 @@ const EditPasswordComponent = ({updateSection}: EditPasswordComponentProps) => {
         }
     };
 
+    // PATCH request to edit Password
+    // Verify input exist & valid or not before submit
     const handleSubmit = async (event: React.FormEvent<HTMLFormElement>) => {
         event.preventDefault(); 
-        try {
-            const response = await fetch ('http://localhost:3000/api/editPassword', {
-                method: 'PATCH',
-                headers: {
-                    'content-type': 'application/json',
-                    'Authorization': `Bearer ${token}`
-                },
-                body: JSON.stringify({userid, oldPassword, newPassword})
-            });
-            const data = await response.json();
-            if (response.ok && data.success) {
-                alert('修改密碼成功，請重新登錄');
-                localStorage.clear();
-                navigate('/');
-            } else if (data.invalidToken) {
-                alert(data.message);
-                localStorage.clear();
-                navigate('/');
-            } else {
-                alert(data.message);
-            }         
-        } catch (errorMsg) {
-            console.error(errorMsg);
-            alert('伺服器發生錯誤，請稍後再試');
+        if (!userid || oldPassword === '' || newPassword === '') {
+            alert('缺少必要欄位');
+        } else if (validateField('oldPassword', oldPassword) !== '' || validateField('newPassword', newPassword) !== '') {
+            alert('請確認格式正確');
+        } else {
+            try {
+                const response = await fetch ('http://localhost:3000/api/editPassword', {
+                    method: 'PATCH',
+                    headers: {
+                        'content-type': 'application/json',
+                        'Authorization': `Bearer ${token}`
+                    },
+                    body: JSON.stringify({userid, oldPassword, newPassword})
+                });
+                const data = await response.json();
+                if (response.ok && data.success) {
+                    alert('修改密碼成功，請重新登錄');
+                    localStorage.clear();
+                    navigate('/');
+                } else if (data.invalidToken) {
+                    alert(data.message);
+                    localStorage.clear();
+                    navigate('/');
+                } else {
+                    alert(data.message);
+                }         
+            } catch (errorMsg) {
+                console.error(errorMsg);
+                alert('伺服器發生錯誤，請稍後再試');
+            }
         }
     }
     return (
@@ -58,9 +68,9 @@ const EditPasswordComponent = ({updateSection}: EditPasswordComponentProps) => {
             <button onClick={() => updateSection('profile')}>回到用戶設定</button>
             <form onSubmit={handleSubmit}>
                 <div>輸入密碼</div>
-                <input type='password' name='oldPassword' value={oldPassword} onChange={OnInputChange} placeholder='舊密碼'></input>
+                <input type='password' name='oldPassword' value={oldPassword} onChange={OnInputChange} placeholder='舊密碼' required></input>
                 <div>輸入新密碼</div>
-                <input type='password' name='newPassword' value={newPassword} onChange={OnInputChange} placeholder='新密碼'></input>
+                <input type='password' name='newPassword' value={newPassword} onChange={OnInputChange} placeholder='新密碼' required></input>
                 <button>提交</button>
             </form>
         </>
